Split text-quote into two files
